+++ astroquery/mast/tests/test_mast.py
@@ -15,6 +15,7 @@ from astropy.io import fits
 import astropy.units as u
 from requests import HTTPError, Response
 
+from astropy.utils.exceptions import AstropyDeprecationWarning
 from astroquery.mast.services import _json_to_table
 from astroquery.utils.mocks import MockResponse
 from astroquery.exceptions import (InvalidQueryError, InputWarning, MaxResultsWarning, NoResultsWarning,
@@ -366,15 +367,51 @@ def test_missions_get_unique_product_list(patch_post, caplog):
 def test_missions_filter_products(patch_post):
     # Filter products list by column
     products = mast.MastMissions.get_product_list('Z14Z0104T')
-    filtered = mast.MastMissions.filter_products(products,
-                                                 category='CALIBRATED')
+    filtered = mast.MastMissions.filter_products(products, category='CALIBRATED')
     assert isinstance(filtered, Table)
     assert all(filtered['category'] == 'CALIBRATED')
 
-    # Filter by non-existing column
-    with pytest.warns(InputWarning):
-        mast.MastMissions.filter_products(products,
-                                          invalid=True)
+    # Filter by extension
+    filtered = mast.MastMissions.filter_products(products, extension='fits')
+    assert len(filtered) > 0
+
+    # Numeric filtering
+    # Single integer value
+    filtered = mast.MastMissions.filter_products(products, size=11520)
+    assert all(filtered['size'] == 11520)
+
+    # Single string value
+    filtered = mast.MastMissions.filter_products(products, size='11520')
+    assert all(filtered['size'] == 11520)
+
+    # Comparison operators
+    filtered = mast.MastMissions.filter_products(products, size='<15000')
+    assert all(filtered['size'] < 15000)
+
+    filtered = mast.MastMissions.filter_products(products, size='>15000')
+    assert all(filtered['size'] > 15000)
+
+    filtered = mast.MastMissions.filter_products(products, size='>=14400')
+    assert all(filtered['size'] >= 14400)
+
+    filtered = mast.MastMissions.filter_products(products, size='<=14400')
+    assert all(filtered['size'] <= 14400)
+
+    # Range operator
+    filtered = mast.MastMissions.filter_products(products, size='14400..17280')
+    assert all((filtered['size'] >= 14400) & (filtered['size'] <= 17280))
+
+    # List of expressions
+    filtered = mast.MastMissions.filter_products(products, size=[14400, '>20000'])
+    assert all((filtered['size'] == 14400) | (filtered['size'] > 20000))
+
+    with pytest.raises(InvalidQueryError, match="Could not parse numeric filter 'invalid' for column 'size'"):
+        # Invalid filter value
+        mast.MastMissions.filter_products(products, size='invalid')
+
+    # Error when filtering by non-existing column
+    with pytest.raises(InvalidQueryError, match="Column 'non_existing' not found in product table."):
+        mast.MastMissions.filter_products(products, non_existing='value')
 
 
 def test_missions_download_products(patch_post, tmp_path):
@@ -670,11 +707,31 @@ def test_observations_get_product_list(patch_post):
 
 def test_observations_filter_products(patch_post):
     products = mast.Observations.get_product_list('2003738726')
-    result = mast.Observations.filter_products(products,
-                                               productType=["SCIENCE"],
-                                               mrp_only=False)
-    assert isinstance(result, Table)
-    assert len(result) == 7
+    filtered = mast.Observations.filter_products(products,
+                                                 productType=["SCIENCE"],
+                                                 mrp_only=False)
+    assert isinstance(filtered, Table)
+    assert len(filtered) == 7
+
+    # Filter for minimum recommended products
+    filtered = mast.Observations.filter_products(products, mrp_only=True)
+    assert all(filtered['productGroupDescription'] == 'Minimum Recommended Products')
+
+    # Filter by extension
+    filtered = mast.Observations.filter_products(products, extension='fits')
+    assert len(filtered) > 0
+
+    # Numeric filtering
+    filtered = mast.Observations.filter_products(products, size='<50000')
+    assert all(filtered['size'] < 50000)
+
+    # Numeric filter that cannot be parsed
+    with pytest.raises(InvalidQueryError, match="Could not parse numeric filter 'invalid' for column 'size'"):
+        filtered = mast.Observations.filter_products(products, size='invalid')
+
+    # Filter by non-existing column
+    with pytest.raises(InvalidQueryError, match="Column 'invalid' not found in product table."):
+        mast.Observations.filter_products(products, invalid=True)
 
 
 def test_observations_download_products(patch_post, tmpdir):
@@ -702,8 +759,7 @@ def test_observations_download_products(patch_post, tmpdir):
 
     # passing row product
     products = mast.Observations.get_product_list('2003738726')
-    result1 = mast.Observations.download_products(products[0],
-                                                  download_dir=str(tmpdir))
+    result1 = mast.Observations.download_products(products[0], download_dir=str(tmpdir))
     assert isinstance(result1, Table)
 
 
@@ -1012,7 +1068,8 @@ def test_tesscut_get_sector(patch_post):
 
     # Exercising the search by moving target
     sector_table = mast.Tesscut.get_sectors(objectname="Ceres",
-                                            moving_target=True)
+                                            moving_target=True,
+                                            mt_type='small_body')
     assert isinstance(sector_table, Table)
     assert len(sector_table) == 1
     assert sector_table['sectorName'][0] == "tess-s0001-1-3"
@@ -1020,26 +1077,31 @@ def test_tesscut_get_sector(patch_post):
     assert sector_table['camera'][0] == 1
     assert sector_table['ccd'][0] == 3
 
+    # Invalid queries
     # Testing catch for multiple designators'
     error_str = ("Only one of moving_target and coordinates may be specified. "
                  "Please remove coordinates if using moving_target and objectname.")
-
     with pytest.raises(InvalidQueryError) as invalid_query:
         mast.Tesscut.get_sectors(objectname='Ceres', moving_target=True, coordinates=coord)
     assert error_str in str(invalid_query.value)
 
-    # Testing invalid queries
-    with pytest.raises(InvalidQueryError) as invalid_query:
-        mast.Tesscut.get_sectors(objectname="M101", product="spooc")
-    assert "Input product must either be SPOC or TICA." in str(invalid_query.value)
+    # Error when no object name with moving target
+    with pytest.raises(InvalidQueryError, match='Please specify the object name or ID'):
+        mast.Tesscut.get_sectors(moving_target=True)
 
+    # Error when both object name and coordinates are specified
+    with pytest.raises(InvalidQueryError, match='Please remove objectname if using coordinates'):
+        mast.Tesscut.get_sectors(objectname='Ceres', coordinates=coord)
+
+    # Testing invalid queries
+    # Invalid product type
     with pytest.raises(InvalidQueryError) as invalid_query:
-        mast.Tesscut.get_sectors(objectname="M101", product="TICA", moving_target=True)
-    assert "Only SPOC is available for moving targets queries." in str(invalid_query.value)
+        with pytest.warns(AstropyDeprecationWarning, match="Tesscut no longer supports"):
+            mast.Tesscut.get_sectors(objectname="M101", product="spooc")
+    assert "Input product must be SPOC." in str(invalid_query.value)
 
 
 def test_tesscut_download_cutouts(patch_post, tmpdir):
-
     coord = SkyCoord(107.27, -70.0, unit="deg")
 
     # Testing with inflate
@@ -1067,6 +1129,8 @@ def test_tesscut_download_cutouts(patch_post, tmpdir):
     # Exercising the search by moving target
     manifest = mast.Tesscut.download_cutouts(objectname="Eleonora",
                                              moving_target=True,
+                                             mt_type='small_body',
+                                             sector=1,
                                              size=5,
                                              path=str(tmpdir))
     assert isinstance(manifest, Table)
@@ -1088,16 +1152,12 @@ def test_tesscut_download_cutouts(patch_post, tmpdir):
 
     # Testing invalid queries
     with pytest.raises(InvalidQueryError) as invalid_query:
-        mast.Tesscut.download_cutouts(objectname="M101", product="spooc")
-    assert "Input product must either be SPOC or TICA." in str(invalid_query.value)
-
-    with pytest.raises(InvalidQueryError) as invalid_query:
-        mast.Tesscut.download_cutouts(objectname="M101", product="TICA", moving_target=True)
-    assert "Only SPOC is available for moving targets queries." in str(invalid_query.value)
+        with pytest.warns(AstropyDeprecationWarning, match="Tesscut no longer supports"):
+            mast.Tesscut.download_cutouts(objectname="M101", product="spooc")
+    assert "Input product must be SPOC." in str(invalid_query.value)
 
 
 def test_tesscut_get_cutouts(patch_post, tmpdir):
-
     coord = SkyCoord(107.27, -70.0, unit="deg")
     cutout_hdus_list = mast.Tesscut.get_cutouts(coordinates=coord, size=5)
     assert isinstance(cutout_hdus_list, list)
@@ -1113,6 +1173,8 @@ def test_tesscut_get_cutouts(patch_post, tmpdir):
     # Exercising the search by object name
     cutout_hdus_list = mast.Tesscut.get_cutouts(objectname='Eleonora',
                                                 moving_target=True,
+                                                mt_type='small_body',
+                                                sector=1,
                                                 size=5)
     assert isinstance(cutout_hdus_list, list)
     assert len(cutout_hdus_list) == 1
@@ -1131,12 +1193,9 @@ def test_tesscut_get_cutouts(patch_post, tmpdir):
 
     # Testing invalid queries
     with pytest.raises(InvalidQueryError) as invalid_query:
-        mast.Tesscut.get_cutouts(objectname="M101", product="spooc")
-    assert "Input product must either be SPOC or TICA." in str(invalid_query.value)
-
-    with pytest.raises(InvalidQueryError) as invalid_query:
-        mast.Tesscut.get_cutouts(objectname="M101", product="TICA", moving_target=True)
-    assert "Only SPOC is available for moving targets queries." in str(invalid_query.value)
+        with pytest.warns(AstropyDeprecationWarning, match="Tesscut no longer supports"):
+            mast.Tesscut.get_cutouts(objectname="M101", product="spooc")
+    assert "Input product must be SPOC." in str(invalid_query.value)
 
 
 ######################